@@ -16,10 +16,11 @@
 package com.kymjs.themvp.view;
 
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.appcompat.widget.Toolbar;
 
 /**
  * View delegate base class
